chore(api,infra): moved internal embed endpoints to /internal; using private listener
