+++ api/embed/api.py
@@ -10,7 +10,7 @@ from ceramic_cache.api.schema import (
     CacheStampPayload,
     GetStampsWithV2ScoreResponse,
 )
-from ceramic_cache.api.v1 import handle_add_stamps_only, handle_get_scorer_weights
+from ceramic_cache.api.v1 import handle_add_stamps_only
 from ceramic_cache.models import CeramicCache
 from registry.api.schema import (
     ErrorMessageResponse,
@@ -47,24 +47,7 @@ class AddStampsPayload(Schema):
     stamps: List[Any]
 
 
-# TODO: check authentication for these endpoints. Ideally the embed service requires an internal API key, but the lambda one is only exposed on internal LB
-@api_router.post(
-    "/stamps/{str:address}",
-    auth=internal_api_key,
-    response={
-        200: GetStampsWithV2ScoreResponse,
-        401: ErrorMessageResponse,
-        400: ErrorMessageResponse,
-        404: ErrorMessageResponse,
-    },
-    summary="Add Stamps and get the new score",
-)
-def add_stamps(
-    request, address: str, payload: AddStampsPayload
-) -> GetStampsWithV2ScoreResponse:
-    return handle_embed_add_stamps(address, payload)
-
-
+# Endpoint for this defined in internal module
 def handle_embed_add_stamps(
     address: str, payload: AddStampsPayload
 ) -> GetStampsWithV2ScoreResponse:
@@ -119,16 +102,3 @@ def validate_api_key(request) -> AccountAPIKeySchema:
     This API is intended to be used in the embed service in the passport repo
     """
     return AccountAPIKeySchema.from_orm(request.api_key)
-
-
-@api_router.get("/weights", response=Dict[str, float])
-def get_embed_weights(request, community_id: Optional[str] = None) -> Dict[str, float]:
-    """
-    ---
-    get:
-      description: Get embed weights
-      responses:
-        200:
-          description: Returns the embed weights
-    """
-    return handle_get_scorer_weights(community_id)
